Separate out Windows-specific calls (P/Invoke to kernel32, ntdll, gdi32, etc) into *.Win.cs files and reintroduce some #ifs.
Currently these are only compiled if the target is .NET 4.7.2 - this isn't ideal since these calls will work fine under .NET Core on Windows. Need a better way to handle this.
