+++ src/FreeImage.Standard/FreeImageStaticImports.Win.cs
@@ -1,0 +1,146 @@
+﻿#if NET472 // TODO: This is Windows specific (P/Invoke), not a .NET FX / Core issue
+using System;
+using System.Runtime.InteropServices;
+
+namespace FreeImageAPI
+{
+	public static partial class FreeImage
+	{
+		/// <summary>
+		/// Retrieves a handle to a display device context (DC) for the client area of a specified window
+		/// or for the entire screen. You can use the returned handle in subsequent GDI functions to draw in the DC.
+		/// </summary>
+		/// <param name="hWnd">Handle to the window whose DC is to be retrieved.
+		/// If this value is IntPtr.Zero, GetDC retrieves the DC for the entire screen. </param>
+		/// <returns>If the function succeeds, the return value is a handle to the DC for the specified window's client area.
+		/// If the function fails, the return value is NULL.</returns>
+		[DllImport("user32.dll")]
+		private static extern IntPtr GetDC(IntPtr hWnd);
+
+		/// <summary>
+		/// Releases a device context (DC), freeing it for use by other applications.
+		/// The effect of the ReleaseDC function depends on the type of DC. It frees only common and window DCs.
+		/// It has no effect on class or private DCs.
+		/// </summary>
+		/// <param name="hWnd">Handle to the window whose DC is to be released.</param>
+		/// <param name="hDC">Handle to the DC to be released.</param>
+		/// <returns>Returns true on success, false on failure.</returns>
+		[DllImport("user32.dll")]
+		private static extern bool ReleaseDC(IntPtr hWnd, IntPtr hDC);
+
+		/// <summary>
+		/// Creates a DIB that applications can write to directly.
+		/// The function gives you a pointer to the location of the bitmap bit values.
+		/// You can supply a handle to a file-mapping object that the function will use to create the bitmap,
+		/// or you can let the system allocate the memory for the bitmap.
+		/// </summary>
+		/// <param name="hdc">Handle to a device context.</param>
+		/// <param name="pbmi">Pointer to a BITMAPINFO structure that specifies various attributes of the DIB,
+		/// including the bitmap dimensions and colors.</param>
+		/// <param name="iUsage">Specifies the type of data contained in the bmiColors array member of the BITMAPINFO structure
+		/// pointed to by pbmi (either logical palette indexes or literal RGB values).</param>
+		/// <param name="ppvBits">Pointer to a variable that receives a pointer to the location of the DIB bit values.</param>
+		/// <param name="hSection">Handle to a file-mapping object that the function will use to create the DIB.
+		/// This parameter can be NULL.</param>
+		/// <param name="dwOffset">Specifies the offset from the beginning of the file-mapping object referenced by hSection
+		/// where storage for the bitmap bit values is to begin. This value is ignored if hSection is NULL.</param>
+		/// <returns>If the function succeeds, the return value is a handle to the newly created DIB,
+		/// and *ppvBits points to the bitmap bit values. If the function fails, the return value is NULL, and *ppvBits is NULL.</returns>
+		[DllImport("gdi32.dll")]
+		private static extern IntPtr CreateDIBSection(
+			IntPtr hdc,
+			[In] IntPtr pbmi,
+			uint iUsage,
+			out IntPtr ppvBits,
+			IntPtr hSection,
+			uint dwOffset);
+
+		/// <summary>
+		/// Deletes a logical pen, brush, font, bitmap, region, or palette, freeing all system resources associated with the object.
+		/// After the object is deleted, the specified handle is no longer valid.
+		/// </summary>
+		/// <param name="hObject">Handle to a logical pen, brush, font, bitmap, region, or palette.</param>
+		/// <returns>Returns true on success, false on failure.</returns>
+		[DllImport("gdi32.dll")]
+		private static extern bool DeleteObject(IntPtr hObject);
+
+		/// <summary>
+		/// Creates a compatible bitmap (DDB) from a DIB and, optionally, sets the bitmap bits.
+		/// </summary>
+		/// <param name="hdc">Handle to a device context.</param>
+		/// <param name="lpbmih">Pointer to a bitmap information header structure.</param>
+		/// <param name="fdwInit">Specifies how the system initializes the bitmap bits - (use 4).</param>
+		/// <param name="lpbInit">Pointer to an array of bytes containing the initial bitmap data.</param>
+		/// <param name="lpbmi">Pointer to a BITMAPINFO structure that describes the dimensions
+		/// and color format of the array pointed to by the lpbInit parameter.</param>
+		/// <param name="fuUsage">Specifies whether the bmiColors member of the BITMAPINFO structure
+		/// was initialized - (use 0).</param>
+		/// <returns>Handle to a DIB or null on failure.</returns>
+		[DllImport("gdi32.dll")]
+		private static extern IntPtr CreateDIBitmap(
+			IntPtr hdc,
+			IntPtr lpbmih,
+			uint fdwInit,
+			IntPtr lpbInit,
+			IntPtr lpbmi,
+			uint fuUsage);
+
+		/// <summary>
+		/// Retrieves information for the specified graphics object.
+		/// </summary>
+		/// <param name="hgdiobj">Handle to the graphics object of interest.</param>
+		/// <param name="cbBuffer">Specifies the number of bytes of information to
+		/// be written to the buffer.</param>
+		/// <param name="lpvObject">Pointer to a buffer that receives the information
+		/// about the specified graphics object.</param>
+		/// <returns>0 on failure.</returns>
+		[DllImport("gdi32.dll")]
+		private static extern int GetObject(IntPtr hgdiobj, int cbBuffer, IntPtr lpvObject);
+
+		/// <summary>
+		/// Retrieves the bits of the specified compatible bitmap and copies them into a buffer
+		/// as a DIB using the specified format.
+		/// </summary>
+		/// <param name="hdc">Handle to the device context.</param>
+		/// <param name="hbmp">Handle to the bitmap. This must be a compatible bitmap (DDB).</param>
+		/// <param name="uStartScan">Specifies the first scan line to retrieve.</param>
+		/// <param name="cScanLines">Specifies the number of scan lines to retrieve.</param>
+		/// <param name="lpvBits">Pointer to a buffer to receive the bitmap data.</param>
+		/// <param name="lpbmi">Pointer to a BITMAPINFO structure that specifies the desired
+		/// format for the DIB data.</param>
+		/// <param name="uUsage">Specifies the format of the bmiColors member of the
+		/// BITMAPINFO structure - (use 0).</param>
+		/// <returns>0 on failure.</returns>
+		[DllImport("gdi32.dll")]
+		private static extern unsafe int GetDIBits(
+			IntPtr hdc,
+			IntPtr hbmp,
+			uint uStartScan,
+			uint cScanLines,
+			IntPtr lpvBits,
+			IntPtr lpbmi,
+			uint uUsage);
+
+		/// <summary>
+		/// Moves a block of memory from one location to another.
+		/// </summary>
+		/// <param name="dst">Pointer to the starting address of the move destination.</param>
+		/// <param name="src">Pointer to the starting address of the block of memory to be moved.</param>
+		/// <param name="size">Size of the block of memory to move, in bytes.</param>
+		[DllImport("Kernel32.dll", EntryPoint = "RtlMoveMemory", SetLastError = false)]
+		private static unsafe extern void MoveMemory(void* dst, void* src, uint size);
+
+		/// <summary>
+		/// The RtlCompareMemory routine compares blocks of memory
+		/// and returns the number of bytes that are equivalent.
+		/// </summary>
+		/// <param name="buf1">A pointer to a block of memory to compare.</param>
+		/// <param name="buf2">A pointer to a block of memory to compare.</param>
+		/// <param name="count">Specifies the number of bytes to be compared.</param>
+		/// <returns>RtlCompareMemory returns the number of bytes that compare as equal.
+		/// If all bytes compare as equal, the input Length is returned.</returns>
+		[DllImport("ntdll.dll", EntryPoint = "RtlCompareMemory", SetLastError = false)]
+		private static unsafe extern uint RtlCompareMemory(void* buf1, void* buf2, uint count);
+	}
+}
+#endif
